# Fix `ascii` and `utf8` and some ccall usage