--- a/base/error.jl
+++ b/base/error.jl
@@ -21,8 +21,8 @@
 error(s::AbstractString) = throw(Main.Base.ErrorException(s))
 error(s...) = throw(Main.Base.ErrorException(Main.Base.string(s...)))
 
-rethrow() = ccall(:jl_rethrow, Void, ())::Bottom
-rethrow(e) = ccall(:jl_rethrow_other, Void, (Any,), e)::Bottom
+rethrow() = ccall(:jl_rethrow, Bottom, ())
+rethrow(e) = ccall(:jl_rethrow_other, Bottom, (Any,), e)
 backtrace() = ccall(:jl_backtrace_from_here, Array{Ptr{Void},1}, (Int32,), false)
 catch_backtrace() = ccall(:jl_get_backtrace, Array{Ptr{Void},1}, ())