--- a/base/pointer.jl
+++ b/base/pointer.jl
@@ -29,7 +29,10 @@ unsafe_convert{T}(::Type{Ptr{T}}, a::Array{T}) = ccall(:jl_array_ptr, Ptr{T}, (A
 unsafe_convert(::Type{Ptr{Void}}, a::Array) = ccall(:jl_array_ptr, Ptr{Void}, (Any,), a)
 
 # unsafe pointer to array conversions
-pointer_to_array(p, d::Integer, own=false) = pointer_to_array(p, (d,), own)
+function pointer_to_array{T}(p::Ptr{T}, d::Integer, own::Bool=false)
+    ccall(:jl_ptr_to_array_1d, Vector{T},
+          (Any, Ptr{Void}, Csize_t, Cint), Array{T,1}, p, d, own)
+end
 function pointer_to_array{T,N}(p::Ptr{T}, dims::NTuple{N,Int}, own::Bool=false)
     ccall(:jl_ptr_to_array, Array{T,N}, (Any, Ptr{Void}, Any, Int32),
           Array{T,N}, p, dims, own)