--- a/base/strings/basic.jl
+++ b/base/strings/basic.jl
@@ -11,16 +11,17 @@ string() = ""
 string(s::AbstractString) = s
 
 bytestring() = ""
-bytestring(s::Vector{UInt8}) = bytestring(pointer(s),length(s))
+bytestring(s::Vector{UInt8}) =
+    ccall(:jl_pchar_to_string, Ref{ByteString}, (Ptr{UInt8},Int), s, length(s))
 
 function bytestring(p::Union{Ptr{UInt8},Ptr{Int8}})
-    p == C_NULL ? throw(ArgumentError("cannot convert NULL to string")) :
+    p == C_NULL && throw(ArgumentError("cannot convert NULL to string"))
     ccall(:jl_cstr_to_string, Ref{ByteString}, (Cstring,), p)
 end
 bytestring(s::Cstring) = bytestring(convert(Ptr{UInt8}, s))
 
 function bytestring(p::Union{Ptr{UInt8},Ptr{Int8}},len::Integer)
-    p == C_NULL ? throw(ArgumentError("cannot convert NULL to string")) :
+    p == C_NULL && throw(ArgumentError("cannot convert NULL to string"))
     ccall(:jl_pchar_to_string, Ref{ByteString}, (Ptr{UInt8},Int), p, len)
 end
 
@@ -269,4 +270,3 @@ function filter(f, s::AbstractString)
     end
     takebuf_string(out)
 end
-