1. Make sure `utf8`, `ascii` and `bytestring` always check if the pointer is `NULL`.
2. Avoid useless copy and dynamic dispatch in `ascii` and `utf8`.
3. Fix some out-of-date `ccall` usage to reduce avoid a few type assertions.