@yuyichao 2016-04-05
..... Does @github not support force pushing to update a pull request anymore??